app/requirements.txt
requests
Pillow
face_recognition
urllib3
uvicorn
fastapi